refactor: extract tab constants
