+++ tabs/constants.py
@@ -1,0 +1,155 @@
+# -*- coding: utf-8 -*-
+"""Константы для вкладок приложения."""
+
+
+def sort_options(options):
+    first = ["Нет"] if "Нет" in options else []
+    middle = sorted([opt for opt in options if opt not in ("Нет", "Другое")])
+    last = ["Другое"] if "Другое" in options else []
+    return first + middle + last
+
+
+STRESS_UNITS = sort_options(
+    [
+        "Па",
+        "кПа",
+        "МПа",
+        "Н/мм²",
+        "Н/см²",
+        "Н/м²",
+        "кН/мм²",
+        "кН/см²",
+        "кН/м²",
+        "МН/мм²",
+        "МН/см²",
+        "МН/м²",
+        "кгс/мм²",
+        "кгс/см²",
+        "кгс/м²",
+        "тс/см²",
+        "тс/м²",
+    ]
+)
+
+UNITS_MAPPING = {
+    "Время": sort_options(["мс", "с", "мин", "ч"]),
+    "Перемещение по X": sort_options(["мм", "см", "м"]),
+    "Перемещение по Y": sort_options(["мм", "см", "м"]),
+    "Перемещение по Z": sort_options(["мм", "см", "м"]),
+    "Удлинение": sort_options(["мм", "см", "м"]),
+    "Удлинение по X": sort_options(["мм", "см", "м"]),
+    "Удлинение по Y": sort_options(["мм", "см", "м"]),
+    "Удлинение по Z": sort_options(["мм", "см", "м"]),
+    "Деформация": sort_options(["—", "%"]),
+    "Пластическая деформация": sort_options(["—", "%"]),
+    "Сила": sort_options(["мН", "Н", "кН", "кгс", "тс"]),
+    "Продольная сила": sort_options(["мН", "Н", "кН", "кгс", "тс"]),
+    "Поперечная сила": sort_options(["мН", "Н", "кН", "кгс", "тс"]),
+    "Поперечная сила по Y": sort_options(["мН", "Н", "кН", "кгс", "тс"]),
+    "Поперечная сила по Z": sort_options(["мН", "Н", "кН", "кгс", "тс"]),
+    "Масса": sort_options(["г", "кг", "т"]),
+    "Напряжение": STRESS_UNITS,
+    "Интенсивность напряжений": STRESS_UNITS,
+    "Нормальное напряжение X": STRESS_UNITS,
+    "Нормальное напряжение Y": STRESS_UNITS,
+    "Нормальное напряжение Z": STRESS_UNITS,
+    "Касательное напряжение XY": STRESS_UNITS,
+    "Касательное напряжение YX": STRESS_UNITS,
+    "Касательное напряжение YZ": STRESS_UNITS,
+    "Касательное напряжение ZY": STRESS_UNITS,
+    "Касательное напряжение ZX": STRESS_UNITS,
+    "Касательное напряжение XZ": STRESS_UNITS,
+    "Крутящий момент Mx": sort_options(["Н·м", "кН·м"]),
+    "Изгибающий момент Mx": sort_options(["Н·м", "кН·м"]),
+    "Изгибающий момент My": sort_options(["Н·м", "кН·м"]),
+    "Изгибающий момент Mz": sort_options(["Н·м", "кН·м"]),
+    "Частота 1": sort_options(["Гц", "кГц"]),
+    "Частота 2": sort_options(["Гц", "кГц"]),
+    "Частота 3": sort_options(["Гц", "кГц"]),
+    "Другое": [],
+}
+
+DEFAULT_UNITS = {
+    "Время": "с",
+    "Перемещение по X": "м",
+    "Перемещение по Y": "м",
+    "Перемещение по Z": "м",
+    "Удлинение": "м",
+    "Удлинение по X": "м",
+    "Удлинение по Y": "м",
+    "Удлинение по Z": "м",
+    "Деформация": "—",
+    "Пластическая деформация": "—",
+    "Сила": "Н",
+    "Продольная сила": "Н",
+    "Поперечная сила": "Н",
+    "Поперечная сила по Y": "Н",
+    "Поперечная сила по Z": "Н",
+    "Масса": "кг",
+    "Напряжение": "Па",
+    "Интенсивность напряжений": "Па",
+    "Нормальное напряжение X": "Па",
+    "Нормальное напряжение Y": "Па",
+    "Нормальное напряжение Z": "Па",
+    "Касательное напряжение XY": "Па",
+    "Касательное напряжение YX": "Па",
+    "Касательное напряжение YZ": "Па",
+    "Касательное напряжение ZY": "Па",
+    "Касательное напряжение ZX": "Па",
+    "Касательное напряжение XZ": "Па",
+    "Крутящий момент Mx": "Н·м",
+    "Изгибающий момент Mx": "Н·м",
+    "Изгибающий момент My": "Н·м",
+    "Изгибающий момент Mz": "Н·м",
+    "Частота 1": "Гц",
+    "Частота 2": "Гц",
+    "Частота 3": "Гц",
+}
+
+PHYSICAL_QUANTITIES = sort_options(
+    [
+        "Нет",
+        "Время",
+        "Деформация",
+        "Изгибающий момент Mx",
+        "Изгибающий момент My",
+        "Изгибающий момент Mz",
+        "Интенсивность напряжений",
+        "Касательное напряжение XY",
+        "Касательное напряжение XZ",
+        "Касательное напряжение YX",
+        "Касательное напряжение YZ",
+        "Касательное напряжение ZX",
+        "Касательное напряжение ZY",
+        "Крутящий момент Mx",
+        "Масса",
+        "Напряжение",
+        "Нормальное напряжение X",
+        "Нормальное напряжение Y",
+        "Нормальное напряжение Z",
+        "Перемещение по X",
+        "Перемещение по Y",
+        "Перемещение по Z",
+        "Пластическая деформация",
+        "Поперечная сила",
+        "Поперечная сила по Y",
+        "Поперечная сила по Z",
+        "Продольная сила",
+        "Сила",
+        "Удлинение",
+        "Удлинение по X",
+        "Удлинение по Y",
+        "Удлинение по Z",
+        "Частота 1",
+        "Частота 2",
+        "Частота 3",
+        "Другое",
+    ]
+)
+
+__all__ = [
+    "STRESS_UNITS",
+    "UNITS_MAPPING",
+    "DEFAULT_UNITS",
+    "PHYSICAL_QUANTITIES",
+]
